+++ R/methods.R
@@ -68,9 +68,7 @@
 }
 
 .rohrbach <- function(taints, conf.level, n, alternative, N.units = NULL, r.delta) {
-  if (is.null(N.units)) {
-    stop("'N.units' missing for evaluation")
-  }
+  stopifnot("missing value for 'N.units'" = !is.null(N.units))
   w <- 1 - taints
   mu <- mean(taints)
   vars <- sum(w^2) / n - (2 - (r.delta / n)) * ((1 / 2) * ((sum(w^2) / n) - stats::var(w)))
@@ -90,10 +88,8 @@
   return(result)
 }
 
-.moment <- function(taints, conf.level, n, alternative, m.type) {
-  if (!(m.type %in% c("inventory", "accounts"))) {
-    stop("Specify a valid population type. Either inventory or accounts.")
-  }
+.moment <- function(taints, conf.level, n, alternative, m.type = c("inventory", "accounts")) {
+  m.type <- match.arg(m.type)
   tall <- subset(taints, taints != 0)
   if (m.type == "inventory" & length(tall) > 0) {
     tstar <- 0.81 * (1 - 0.667 * tanh(10 * abs(mean(tall))))
@@ -173,12 +169,10 @@
 }
 
 .direct <- function(bookvalues, auditvalues, conf.level, alternative, N.items = NULL, n, N.units = NULL, correction = FALSE) {
-  if (is.null(N.items)) {
-    stop("missing value for 'N.items'")
-  }
-  if (is.null(N.units)) {
-    stop("missing value for 'N.units'")
-  }
+  stopifnot(
+    "missing value for 'N.items'" = !is.null(N.items),
+    "missing value for 'N.units'" = !is.null(N.units)
+  )
   w <- mean(auditvalues)
   s <- stats::sd(auditvalues)
   tVal <- if (alternative == "two.sided") stats::qt(p = conf.level + ((1 - conf.level) / 2), df = n - 1) else stats::qt(p = conf.level, df = n - 1)
@@ -197,9 +191,7 @@
 }
 
 .difference <- function(bookvalues, auditvalues, conf.level, alternative, N.items = NULL, n, correction = FALSE) {
-  if (is.null(N.items)) {
-    stop("missing value for 'N.items'")
-  }
+  stopifnot("missing value for 'N.items'" = !is.null(N.items))
   we <- mean(bookvalues - auditvalues)
   s <- stats::sd(bookvalues - auditvalues)
   tVal <- if (alternative == "two.sided") stats::qt(p = conf.level + ((1 - conf.level) / 2), df = n - 1) else stats::qt(p = conf.level, df = n - 1)
@@ -218,9 +210,7 @@
 }
 
 .quotient <- function(bookvalues, auditvalues, conf.level, alternative, N.items = NULL, n, correction = FALSE) {
-  if (is.null(N.items)) {
-    stop("missing value for 'N.items'")
-  }
+  stopifnot("missing value for 'N.items'" = !is.null(N.items))
   w <- mean(auditvalues)
   sw <- stats::sd(auditvalues)
   b <- mean(bookvalues)
@@ -244,12 +234,10 @@
 }
 
 .regression <- function(bookvalues, auditvalues, conf.level, alternative, N.items = NULL, n, N.units = NULL, correction = FALSE) {
-  if (is.null(N.items)) {
-    stop("missing value for 'N.items'")
-  }
-  if (is.null(N.units)) {
-    stop("missing value for 'N.units'")
-  }
+  stopifnot(
+    "missing value for 'N.items'" = !is.null(N.items),
+    "missing value for 'N.units'" = !is.null(N.units)
+  )
   w <- mean(auditvalues)
   sw <- stats::sd(auditvalues)
   b <- mean(bookvalues)