+++ tests/testthat/test-consistency-statistical-tables.R
@@ -1053,7 +1053,7 @@ test_that(desc = "(id: f7-v0.5.5-21) Statistical Sampling Results based on the B
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "binomial")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "binomial")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1089,7 +1089,7 @@ test_that(desc = "(id: f7-v0.5.5-22) Statistical Sampling Results based on the B
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "binomial")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "binomial")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1125,7 +1125,7 @@ test_that(desc = "(id: f7-v0.5.5-23) Statistical Sampling Results based on the B
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "binomial")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "binomial")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1161,7 +1161,7 @@ test_that(desc = "(id: f7-v0.5.5-24) Statistical Sampling Results based on the B
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "binomial")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "binomial")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1199,7 +1199,7 @@ test_that(desc = "(id: f7-v0.5.5-25) Statistical Sampling Results based on the P
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "poisson")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "poisson")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1235,7 +1235,7 @@ test_that(desc = "(id: f7-v0.5.5-26) Statistical Sampling Results based on the P
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "poisson")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "poisson")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1271,7 +1271,7 @@ test_that(desc = "(id: f7-v0.5.5-27) Statistical Sampling Results based on the P
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "poisson")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "poisson")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1307,7 +1307,7 @@ test_that(desc = "(id: f7-v0.5.5-28) Statistical Sampling Results based on the P
   tab <- matrix(NA, nrow = length(n), ncol = length(k))
   for (rows in 1:length(n)) {
     for (columns in 1:length(k)) {
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "poisson")
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "poisson")
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1349,7 +1349,7 @@ test_that(desc = "(id: f7-v0.5.5-29) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1389,7 +1389,7 @@ test_that(desc = "(id: f7-v0.5.5-30) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1429,7 +1429,7 @@ test_that(desc = "(id: f7-v0.5.5-31) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1469,7 +1469,7 @@ test_that(desc = "(id: f7-v0.5.5-32) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1509,7 +1509,7 @@ test_that(desc = "(id: f7-v0.5.5-33) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1549,7 +1549,7 @@ test_that(desc = "(id: f7-v0.5.5-34) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1589,7 +1589,7 @@ test_that(desc = "(id: f7-v0.5.5-35) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1629,7 +1629,7 @@ test_that(desc = "(id: f7-v0.5.5-36) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1669,7 +1669,7 @@ test_that(desc = "(id: f7-v0.5.5-37) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1709,7 +1709,7 @@ test_that(desc = "(id: f7-v0.5.5-38) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1749,7 +1749,7 @@ test_that(desc = "(id: f7-v0.5.5-39) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }
@@ -1789,7 +1789,7 @@ test_that(desc = "(id: f7-v0.5.5-40) Statistical Sampling Results based on the H
       if (n[rows] >= N) {
         next
       }
-      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 1, method = "hypergeometric", N.units = N)
+      ss <- jfa::evaluation(conf.level = conf.level, n = n[rows], x = k[columns], materiality = 0.99, method = "hypergeometric", N.units = N)
       tab[rows, columns] <- ceiling(ss$ub * 1000) / 10
     }
   }